Modificado: (testing) [mainwindow] Temperatura.
int->double. Se muestra el valor redondeado con round(), así de esta manera con una rápida visualización se pueden observar las temperaturas de los PMTs.

+++ mainwindow.h
@@ -116,9 +116,9 @@ private:
     void resetHitsValues();
     void getPMTLabelNames();
     void drawTemperatureBoard();
-    void setTemperatureBoard(int temp, QLabel *label_pmt, int pmt);
+    void setTemperatureBoard(double temp, QLabel *label_pmt, int pmt);
     void clearTemperatureBoard();
-    temp_code getTemperatureCode(int temperature);
+    temp_code getTemperatureCode(double temperature);
     void getARPETStatus();
 
     /* Preferencias */
